Fix crash in LightSource sorting
When building the RenderCommands, for each one, the active light source vector
needs to be sorted to find the 8 nearest lights. Since n RenderCommands may be
built at the same time, we should only sort on a copy of the RenderView global
active light source vector as: 1) we don't want to modify the global one 2)
that would result in a race and eventually a crash.

Change-Id: I5ac66b5f6af67876f4e374def53bf2bc5d0ba4b9
Task-number: QTBUG-54634
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>