Lights: Fix animation in example and add another
Looks like QML doesn't like animating only a single element of a 3d
vector. It seems to default the non-animating elements to 0. Instead
use a Vector3DAnimation.

Also added another animation to the transformation to make the
spotlight move back and forth in addition to sweeping it's local
direction angle.

Change-Id: Ief490752107b2b618dcf8500cb87561b3baea0bb
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>