Allow to add/remove attributes imperatively in QML
If one needs to add an attribute to an existing geometry property,
currently the attributes property isn't available in such a situation.
This is especially problematic with Mesh. For now, simply give access to
addAttribute and removeAttribute on the QML side, so that it can be
workarounded imperatively for the time being.

Change-Id: I8f4d675d962f084c1c4e0ab4aecc907ae891f83a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>