Remove qboundingvolumedebug
Wasn't being built anyway which led to a warning with moc. We should
rethink how we do such debug rendering so as to not hardwire a Layer
component and make it work with other framegraphs not implementing a
simple forward renderer. Can revert this in the future as a starting
point if needed.

Change-Id: I055310f440b49f144a5ab635122f2193b8c5ac12
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>