Improve GLTextureManager
By making it generic so that it could be reused by any API specific texture
and texture images. That way the texture sharing could be reused no matter
which renderer is being used.
Also fix an assert on destruction due to the texture calling the cleanup
method where the cleanup method would try to delete something already deleted.

Change-Id: Ifc50644edc4a29931102f6e1ba1277cdbc0d843c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>