Temporarily invert the cubemap images
As a result of 44f1a92338859f1cc043ec22e0142037b5f87936 the cubemap
texture images get mirrored. As we cannot add the api to control this
until Qt 5.8, for now we temporarily mirror the textures on disk. We
should revert this in 5.8 once the mirrored property is in.

Task-number: QTBUG-54881
Change-Id: I675f4dc1576e076c63f8c592cdbc51707afb4cb7
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>