Fix cube geometry to have correct texture coords and tangent vectors
The normal lighting materials were behaving weirdly with cubes. Some
faces were behaving correctly whilst others were inverting the lit and
shadowed sides of normal mapped areas (the slopes). Tracked this down
to some inconsistencies in the generation of the cube geometry.

The cube geometry now generates data consistently such that:

* when you look at each face from the following locations and
  orientations:
** Positive X face with positive y as up vector
** Negative X face with positive y as up vector
** Positive Z face with positive y as up vector
** Negative Z face with positive y as up vector
** Positive Y face with negative z as up vector
** Negative Y face with positive z as up vector

then:

* the u texture coordinate always increases to the right as you look at
  any face of the cube
* the v texture coordinate always increases up the face
* the tangent vector is aligned with the u texture coordinate

and taking special care that the w component is correctly oriented.

This can be tested by using a normal mapped cube with a point light
orbiting around above it takign note of the regions that are lit and
shadowed.

Change-Id: I834749b08efb9e03548e8c749c6281958a4e6329
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>