GraphicsContext: do not reset m_stateSet
The renderer resets to the default state after rendering a frame.
The GraphicsContext resetting the stateset to nullptr internally breaks the
assumption the renderer made that the state was reset to the default state. In
turn this would cause some states to not be properly reset when the first
RenderView of a frame defines a stateset as the delta wouldn't be between the
default stateset and the RV's stateset but between nullptr and the RV's stateset.

Change-Id: I46cdf8983278090b067d2a67ad8a93512718a317
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>