Entity: fix -> properly clear childrenHandles vector
When cleanup is called that means an Entity needs to be recycled for reuse
later on. The children handles vector wasn't cleared which resulted in it
containing invalid handles when the Entity element was reused. In turn that
eventually resulted in asserts in the handle manager when that Entity was
released again (which explains crashes with the QSceneLoader changing
sources).

Change-Id: Ic2e69132e4a5186f92936593ddcf3da72a7e1446
Task-number: QTBUG-54889
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>