QSkyboxEntity: remove camera translation property
If we convert a mat4 matrix to a mat3 matrix, then back to a mat4, this
removes the translation component from it. Making the skybox shaders use that
property allows to remove references to the camera translation in
QSkyboxEntity.

This makes using a skybox neater as we don't need to care about the Camera
position anymore. This also allows to render the same skybox with different
cameras which required to have duplicate QSkyboxEntity previously.

Change-Id: I1adddd17d151e4d26d0f19b048df88dbf738042b
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>