QAbstractTexture::CubeMapFace: add a new entry AllFaces
When rendering a cube map to an FBO, we may want to specify that we want to
render to all the faces rather than a single one. This is indeed possible with
a geometry shader and using gl_Layer. For such a case we need to be able to
specify that in the QRenderTargetOutput face property.

Note: gl_Layer can also be used with texture arrays but a neat api to specify
which layers to use on QRenderTargetOuput seems a bit more tricky in that case.

Change-Id: I6abf8667fd0905d78c22ef8d5f1f792a5973bbbe
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>