GraphicsHelpers: bindFrameBufferAttachment
Handle binding all the faces of a cube map texture if supported. Print a
warning if not supported.

This can be really useful when you want to render a dynamic sky into a cube
map texture to be used for reflection or refraction without having to create 6
FBO or 6 attachments which would require 6 passes to be rendered.

Change-Id: Id5a920d608d895878a177f777f898a367f96aeb1
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>