Add an AbstractActionInput base class
This is needed for the coming refactoring improving testability in the
input aspect. Right now we got some of the action input logic done
inside the job and some inside the class. This is in part due to the
lack of common interface between the different action input backend
node. This is a first step toward solving that.

Change-Id: I803ecd1f114f635d2b275d5dadc493733fc580e8
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>