Renaming Scene IO classes to scene import classes
The so-called "IO" classes have never done I/O, only import.
There is no sense to add also export to those classes as the IF
is designed to do only import and there is no code that would be
shared between the importer and exporter/saver.
For the sanity of the code the naming is changed and we will
implement exporter/saver class to be used for storing purposes.

Change-Id: I832bef0e7de1f65289a89c6898a475dab7431dc0
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>