Add unit test for QTorusGeometry
Also fix up some more problems with the torus. We were drawing some
triangles more than once due to problems in the indexing. Fixed this
by factoring out the calculations of number of vertices and
triangles and takign care about needing more vertices to handle the
texture coordinate wrapping at the seams but not more indices.

Also renamed some members in the functors to be more consistent.

Now the torus geometry should be perfect.

Change-Id: I1198469531a5fd5e4ec8b86d47c9e66850fe50c7
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>