Fix orientation of textures on planes in audio vis example
Due to 44f1a92338859f1cc043ec22e0142037b5f87936 we need to correct
the orientation of the textures applied to the plane meshes in this
example. There is a variety of ways to do this including:

* Adapting the texture coordinates
* Rotate plane other direction about x axis and swap face culling mode
* Flip textures on disk

This patch goes for the 3rd option. This can be reverted in 5.8 when we
get new API to control whether texture images are flipped upon loading.

None of these are ideal but is an intermediate step in getting to a
fully consistent usage with 5.8. This follows the same approach as was
used with the anaglph-rendering example skybox textures.

Change-Id: I7fda41204625d7b30ac295ef789b6dea6018c2b8
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Mika Salmela <mika.salmela@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>