UpdateAxisActionJob: remove misleading redundant check
This code is in the then-branch of an if (buttonInput),
and buttonInput is not modified in between, so the
check for buttonInput here is redundant.

So remove it.

Also declare 'buttonInput' as const to have the above
analysis checked at compile-time, too.

Introduced in 496d8b53b16bd0703ab06978f18fcbbb7d95e845,
so Qt 5.6 is unaffected.

Coverity-Id: 161331
Change-Id: I86e4d81e534d7906281fc36b70dc57ccb56184ca
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>