QRenderAspect: remove misleading check
The check of d->m_renderer against nullptr follows an unconditional
deref of the same pointer with no intervening code. It must therefore
be always true.

Remove it.

That done, m_renderer _is_ nullptr after the ctor ran, and before
the dtor runs (there's a check), so maybe the code should at least
assert the existence of m_renderer before using it.

Coverity-Id: 156307
Change-Id: Iacf2c09db1c0a5a55a67cfb6ef2a00652e557d09
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>