QEntity::createNodeCreationChange fix
Was sending the dynamic metaobject of the QComponent which (in the case of the
QComponent being destroyed while the change hasn't yet been processed) would
result in crashes.
Moved out the findStaticMetaObject function from QNodeCreatedChange and made it
a static function of QNodePrivate. Use QNodePrivate::findStaticMetaObject in
the QEntity::createNodeCreationChange.

Change-Id: I15a63251d7f3aae25191fa4440d864f0c8234a61
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>