GraphicsContext: remove misleading check in specifyAttribute()
The 'location' argument is checked for non-negativeness first thing
in the function, and is not modified in between, so the re-check in
the loop is both unneeded and misleading, as it suggests that the
variable may have changed from the previous test.

Fix by removing the second check.

Introduced by 8c1e1df4be02de8d64807d7abeccedc6becd3076, which means
only 5.7+ is affected.

Coverity-Id: 168546
Change-Id: I3adc0ace2eac234142ab856738e553b1a7e76e1f
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>