NodeManagers: fix invalid template specialization
When we changed the way we handled VAO by creating the OpenGLVertexArrayObject
class (which is a wrapper around QOpenGLVertexArrayObject), we adjusted the
VAOManager to store OpenGLVertexArrayObject instead of
QOpenGLVertexArrayObject pointers, the template specialization was not
properly updated.
This patch fixes that by using in the specialization the type that is actually
being stored by the manager.

Change-Id: I1b5fd20a1dddf4a8429aafdce4cc35488e77a379
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>