NodeInstantiator created nodes go to its parent
Currently, elements created by the NodeInstantiator are children of the
instantiator itself. Change that to parent them in the instantiator
parent instead.

Rationale for that is that it will behave similarly to Repeater this
way, and gives a chance of using NodeInstantiator in frame graphs.
Indeed, anything which is not a FrameGraphNode gets pruned in that
context that would include branches starting from NodeInstantiator.

Change-Id: I651a51471d92fcf466c9abd5dbbbfa2960ef1247
Task-Id: QTBUG-55908
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>