GLTF Scene export plugin
Added a plugin that can export Qt3D entity tree as a
GLTF scene. The GLTF format is extended a little to support
those cases not covered by standard format.
The existing GLTF import plugin was modified to support cases
it didn't already support.

Any meshes, transformations, default materials from Qt3DExtras,
cameras, and lights can be exported and imported.
Other components are ignored.

Adding support for custom materials is left for the future.

The camera fov unit (degrees vs radians) was also fixed to qgltf
tool to make it consistent with both assimp and GLTF specs,
as well as the fixed implementation in the importer plugin.

Change-Id: Iaa0aa6656429abb173f291fb1a768bf98819ac84
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>