Fix picking on child entities
Picking was broken for child entities of an entity having an object
picker. This commit addresses it. It was due to in practice always
testing the entities with object picker and not their children since
when verifying the presence of an object picker the information of the
original entity tested was lost.

Change-Id: Icce29b2bf4424ae8c51a8b4fc26cbe9b9b0041e5
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>