Fix build with ICC during -Werror: unused local variable
This variable is in an anonymous namespace, so ICC rightly finds out
that nothing uses it. Declare our intentions properly in the comment and
add the Q_DECL_UNUSED to tell the compiler to shut up.

error #2415: variable "Qt3DRender::Render::<unnamed>::qNodeIdTypeId" of static storage duration was declared but never referenced

Change-Id: I73fa1e59a4844c43a109fffd148cccdc579c3851
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>