Fix texture data upload timing
Following the texture refactoring, textures are correctly uploaded at
application startup. However if the content of texture data is later
changed, e.g. by changing source property of a TextureImage, this is
never made available to the GPU.

Tracing this through, the texture image data generator is correctly
executed but it seems the OpenGL submission thread tries to upload the
data before the new data is made available to the TextureDataManager.
The subsequent clearing of the dirty flag means it never gets uploaded.

To fix this the TextureData dirty flag in GLTexture has been re-purposed
to mean that the generators have been executed and the data is ready and
available for upload. This flag was actually redundant for the purpose
of scheduling the generators to be executed as that is handled directly
by the texture data manager and GLTexture.

With this commit, the upload is never attempted until after the texture
image data is ready and we see textures behaving correctly once again.

Task-number: QTBUG-57509
Change-Id: I662c8d17e17283a2a3be5afbf61289ec2405d4fc
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>