Work around QML -> C++ float to int conversion
QML converts values of reals that have zero fractional parts e.g. 2.0
to integers when passing them to C++. This can result in Qt 3D and the
Uniform class passing these to uniform floats in GLSL by
reinterpret_casting the integer to a float which results in a garbage
value in the shader which in turn leads to very difficult to debug
shader issues.

This commit handles this particular case (in a hacky way) for the
Qt 5.8.0 release. We need to find a more fully featured solution for
converting Uniform's to the correct introspected GLSL types. This will
require being in position of the introspected shader interface when
building the RenderViews. This is too large of a change to do in time
for the 5.8.0 release. Hence this work around solution for now.

Task-number: QTBUG-57510
Change-Id: I6631879fff8259a32960e461bad4c692ca630220
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>