Add Qt3DExtras::QTextureAtlas class
areallocator is copied from qtdeclarative/src/quick/scenegraph/util.
We could link to QtQuick to use this allocator from the private headers,
but this would introduce a hard dependency on QtQuick that Qt3DExtras
doesn't yet have.

Image data for texture atlasses are stored inside QTextureAtlasData
structures, which are shared between
  1. the frontend QTextureAtlas node, and
  2. the texture generators that are updated each time something is
     added to the atlas

If we didn't share this data, we would have to copy the current QImage
that stores the current state of the texture atlas, each time that a
new sub-image is added. This would result in considerable memory copying
overhead.

By sharing the data, we can just add new sub-images to the shared data
structure, and update the texture generator. This may happen dozens of
times within one frame. When the backend texture loading job is executed,
it will copy all the new sub-images into the overall texture image, and
create the texture data from that image. This way, exactly zero image
copying overhead happens in the frontend thread.

Change-Id: I8c418bf335afd1363ad7cefdf81778e4075038e8
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>