Transfers gpu buffer data to cpu
Added some code to retrieve the data in the gpu back to a cpu pointer
in the opengl helpers.

Added a framegraph node which specifies if any data must be downloaded
from the gpu

Added a flag in the buffer object to specify if the data from that
buffer must be downloaded.

Added a job to send the data back to the frontend.

Added a path in the renderer to download the gpu data.

Task number: QTBUG-57727

Change-Id: I315a9221e5116c7d07f818e00c654fe1901144f4
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>