QDistanceFieldGlyphCache: stores distance field textures
For text rendering via textures (independent of the technique used)
we want to store font glyphs
  1. in a texture atlas, in order to reduce the number of GL textures
     by orders of magnitude
  2. reference-counted: so that we can delete glyphs that are no
     longer needed, i.e. that are no longer rendered by any text
     instantiation

The QDistanceFieldGlyph cache is a node that has to be placed somewhere
in the scene, and is then referenced by QDistanceFieldText nodes so
it will be used for text rendering.

It associates with each QRawFont (but independent of the font size) a
DistanceFieldFont, which stores in one or more texture atlasses (one
atlas will hold ~50 entries, so it will be sufficient for most font
instantiations) the glyphs that are currently referenced.

These texture atlasses will be children of the QDistanceFieldGlyphCache.
The QDistanceFieldGlyphCache offers methods to ref and de-ref glyphs for
a given QRawFont, and will return a struct that contains the texture
the glyphs are stored in, and the associated texture coordinates
within that texture.

Change-Id: I3c9557ecad605e912f8ce3b92265b9fc39329f49
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>