Fix node de-registration from scene on deletion/reparenting
QNodePrivate::unsetSceneHelper() used to be broken, in that it set
the m_scene of the the item that the NodeVisitor functor was executed on.
Subsequent calls for child nodes would then have a m_scene == nullptr,
and could thus not be de-registered.

This patch fixes this behavior by treating each node's m_scene separately

Task-number: QTBUG-57571
Change-Id: I779caadd6957a97514fd105cd16935c3edc9f41f
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>