Add QAnimationClip frontend class and test
Represents an animation clip. To keep the public API minimal just use
a source property for now. API can be extended later to allow the user
to procedurally create animation clips.

Assumption for now is that we can export a suitable format from a
digital content creation tool such as blender.

Change-Id: Ie18b642a92e3cedb1833f013ffe1f7c0f5d07dac
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>