Send further pick events to initial object picker
Once an object picker receives a press event, all further
events will be sent to it until the mouse is released.
This enables interactive with objects that move behind
other pickable objects. In that case, the pick event
will be sent to the active object picker.

This behavior is disabled if all pick events are
requested (which works as before).

Included an example where a cylinder is dragged to
control the radius of a cone (like a slider) and
moves behind another pickable object (while remaining
interactive).

Task-number: QTBUG-55495
Change-Id: I4e3b8466d01ca1d08ef48c1fa6f6abd5b5404980
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>