Keep the existence of renderer and nodes in sync
Create and delete the node manager along with
the renderer. This is require because backend
node have a pointer to the renderer so should
not outlive it.

Note that unregistering the render aspect is
probably not safe since the frame advance service
used in the aspect manager is owned by the renderer
(which gets destroyed when render aspect is
unregistered).
Also re-registering the render aspect does not
cause the backend nodes to be recreated since
startup is not called then.

Change-Id: I9d301d1b49528133bf28e3ae9ee5acf469c449ae
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>