Picking: restore early checks
5.8.0 removed early checks because it broke hover.

Restoring early checks by looking for at least one enabled
object picker that requires hover support.

Added a flag to job that can be set by pickers when they
are created, deleted or some property changed. This is
used to lazily update the state flags for the early
termination checks.

Also removed the dirty flag handling as it was not used
and never cleared.

Task-number: QTBUG-57611
Change-Id: Ib5c60290ee016d8c1fd3f24f7ebacff412fec239
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>