LevelOfDetail (LOD) node
Component to control complexity of entities based on either
distance to camera or size on screen.

Given a specific camera, the LevelOfDetail computes the distance to
the observer, or the screen size of the entity. The currentIndex
property is updated to identify which of the entries in the
provided array of distance or size ranges contains the current value.

A disabled LevelOfDetail component will act as a simple
(user controlled) switch.

Updates are not done on disabled sub trees since their
bounding boxes may not be valid.

Change-Id: Ifa4c088de6b80b6a6ed203070411761574686cdb
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>