Change QTextureLoader defaults
We realized that for rich formats the defaults are ignored by
QTextureLoader. At the same time for simpler formats the defaults coming
from QAbstractTexture are a bit subpar for quality. So change
QTextureLoader defaults to be better in that regard to make
QTextureLoader a nice way to load a texture from file in one line.

Change-Id: Iefe1b07d14f362ef619e4a86d6599370534041ff
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>