Consolidate QML/C++ materials in extras
We remove the qml implementation of materials in extras that were also
implemented in C++. Note this is an API change for the materials having
texture properties. The QML ones had mistakenly a string property
instead of a texture one. This move also fixes this API issue on the QML
end.

Change-Id: Ibed14288cd7b8c5ab9615b74b949c6a73ac29329
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>