Allow Quick3DEntityLoader to work when created from C++
For that we try harder at finding the QML engine. If it's not known to
the Quick3DEntityLoader instance, we climb up the parent/child
relationship to find an object associated to the engine and we get the
loader to work from there.

In practice will be useful for LevelOfDetailLoader in the next commit.

Change-Id: I64ab709d25e8edbecc3662eb96f5d92f99f1b13e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>