Shutdown fixes
Wait for render thread to finish when shutting down. Was causing
crashes in some background nodes still being accessed after the
node manager was destroyed.

Shader disconnects from glcontext destroy messages when that
context is reset. Thread affinity of glcontext and shader meant
the destroyed signal from the glcontext could be delivered to
a shader that had already been destroyed.

Change-Id: I283701e8d008343b3fccf6769d71695b1c6a1ea4
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>