Clean up QRenderCapture(Reply) API
We don't need the captureId in the public API so deprecate the functions
which make it appear and create overloads with no such id when
appropriate. Only a pointer to a QRenderCaptureReply is needed to
represent a capture request.

Also the "isCompleted" parameter is unneeded on the signal since it is
always true anyway. If we'd want to do error management it would go via
a status flag property or similar.

Change-Id: I9571ece3e3f82f46db5b3993ccf035e770c9d55e
Task-Id: QTBUG-58877
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>