Revert "Use QVector<float> in level of detail APIs"
This reverts commit 5bceaee38908934078bbdb62b564daa68f7e182b.

QML uses doubles throughout and there is no good way or place to
reliably switch to floats in the API. We tried extending
QtDeclarative to work with QJSValues containing vectors/lists of
floats and doubles (rather than qreal) but this introduces
rounding errors that are difficult to control robustly.

We also looked at removing the Q_PROPERTY in C++ and adding it
to a QML extension object. But that would make it not work with
animations bound to this property.

So, in the end, it seems like qreal is the best of a bad set
of alternatives. We should make a renewed effort to kill qreal
for Qt 6.

Change-Id: I9d61e58e7223eb5a6b848ba33fc760b3654bbddd
Reviewed-by: Mike Krus <mike.krus@kdab.com>