Add support for line picking
PickingSettings can ask to get line as well as triangle picks.
Introduces a radius value to compensate for numerical precision
in ray-segment intersections.
Introduces QPickLineEvent with the details about the picking.

Job will perform line picking if appropriate. Hit encode the
type of picking and this is used to generate the right type of
event.

Task-number: QTBUG-58071
Change-Id: I834e6cc08044a8cfb28bba7443034e05267aedbf
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>