Rename CoordinateVisitor to BufferVisitor
Point picking will need a point/coordinate visitor
to traverse coordinate in order. BufferVisitor traverses
a single buffer irrespective of the draw order.

It reveals (another, pre-existing) problem with bounding
volume computations: it computes based on all the
values in a buffer. In the case of indexed primitives,
only a subset of those may be used and the bounding
volume may be very different. Will fix this for 5.10
using proper coordinate visitor

Change-Id: Ie86f26d7066260d69da9639eb1dab823c650333c
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>