Use BVH to accelerate picking
Previously, picking gathered the list of all entities and used
QtConcurrent to run pick queries on all of them.

This changes logic so that picking is done hierarchically from the root.
If the bounding volume of a sub-graph does not hit, then it’s culled
entirely. Otherwise, the entity itself is tested picked, and each
of the child sub-graphs.

Task-number: QTBUG-48983
Change-Id: I06cf2db07b1822a16a5cf82fc7416de5274115ef
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>