Add ClipResults to ClipBlendNode indexed by animator id
A blend tree (or subset of) can be shared between multiple animators.
This implies that when evaluating blend trees we need to be able to
store and retrieve the results of node evaluations on a per animator
basis (as each animator may well have a different start time and
therefore map to different local times for the same blend tree).

Presently this is done by storing the blend tree node evaluation
results within the jobs themselves. Instead, let's delegate this to
the blend tree nodes.

This is the first step in making the blend tree more general to support
more complicated blend operations such as generalised lerps,
barycentric lerps, multidimensional lerps etc. Future commits will
build up support for this in parallel with the current implementation
and then remove the existing implementation when complete.

Change-Id: Ife097ec11571c71e9e3d420b3ca67efa2e1639cc
Reviewed-by: Mike Krus <mike.krus@kdab.com>