Add virtual to ClipBlendNode to allow querying dependencies
This will allow a job to build up a list of clips to be evaluated for
the blend tree. This has enough flexibility to allow us to support
many types of bend node in the future but still allow us to only
evaluate the clips actually needed at any given time and for the
current set of blend node states.

Change-Id: I6bfff6d18434af4f09764d890af6f8f429996994
Reviewed-by: Mike Krus <mike.krus@kdab.com>