Fix m_uniformnames Iterator
I'm new to the source, but this seems to be wrong at first glance since it was using the attributes instead of the uniforms. Could you verify the change is what should be intended? It's doubtful that this would be a problem since it checks the m_uniforms collections too, but just to be safe.