Scene2D cleanup - register pick events when scene is initialized
The current model where mouseGrab is connected to picker signal
doesn't work. The pressed event is not sent to scene2d item, because it is
already lost when scene2d gets the grabMouse message where it registers to
the picker events. This breaks the mouse event sequence
(press-move-release) for the quick item.

Instead hook to setScene of the node to message scene2d when the scene gets
initialized and always register to the picker events.

Task-number: QTBUG-58876
Change-Id: Ic9ca4b0899a030336ef20ff2cffbe10b567c36f5
Reviewed-by: Antti Määttä <antti.maatta@qt.io>