Add function to calculate phase of animation from global time
When we come to evaluate animation blend trees, the resulting
duration of the blend tree will also be blended. This means that we
can't just map from global time to local time of each leaf node clip.
Instead we need to evaluate the clips at a consistent phase. This way
we can safely blend a 3 second walk animation cycle with a 2 second
run animation cycle safely. So long as the animator lines up the
foot falls at consistent phases.

We will therefore need to be able to calculate the phase from the
current global time and the blend tree's duration before we can
eventually evaluate the clips at the phase.

Change-Id: I7bce231cdc2eee8cb873f2c98c6c8c2e05533c8c
Reviewed-by: Mike Krus <mike.krus@kdab.com>