Add function to recursively calculate duration of a blend tree
The resulting duration is independent of which animators the blend
tree is associated with so this can be implemented as a simple set
of virtual functions in the concrete classes.

The value node simply returns the duration of the contained clip; the
additive node returns the duration of the base node; and the lerp node
lerps the durations of the start and end nodes.

Change-Id: Ib6edea3fa495885493fa72d44437fea5a8c5a446
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>